+++ .gitignore
@@ -104,3 +104,6 @@ dist
 
 # TernJS port file
 .tern-port
+
+# Local Netlify folder
+.netlify
