+++ README.md
@@ -1,10 +1,16 @@
 # Apotheken Notdienst Crawler
 
+# UPDATE: BLAK changed their website and the crawler is not working anymore. I will not update this project anymore.
+
+The crawler is not working anymore because the website changed because BLAK now provides a free XML API. Register here: https://www.blak.de/notdienst/notdienst-und-dienstbereitschaft/schnittstelle-xml-html
+
+---- 
+
 ### A simple crawler in Node.js to build an API to get the current emergency pharmacies for a given location in Germany directly from [BLAK](http://www.lak-bayern.notdienst-portal.de/blakportal/). I needed that quickly for a small project.
 
 ## Introduction
 
-This repository contains code for a web crawler that scrapes a website. The API can be pubslished using a serverless lambda functions at hosters like [Netlify](https://netlify.com).
+This repository contains code for a web crawler that scrapes a website. The API can be published using a serverless lambda functions at hosting providers like [Netlify](https://netlify.com).
 
 Comes with absolutely no warranty, API is subject to change and it needs refactoring / sanity measures. Use on your own risk.
 
@@ -16,9 +22,9 @@ Install dependencies: `npm install`
 
 This project was built with the following tools:
 
-- Node: A JavaScript runtime built on Chrome's V8 JavaScript engine.
-- Axios: A promised based HTTP client for the browser and Node.js.
-- Cheerio: A lightweight implementation of jQuery which provided access to the DOM on the server.
+* Node: A JavaScript runtime built on Chrome's V8 JavaScript engine.
+* Axios: A promised based HTTP client for the browser and Node.js.
+* Cheerio: A lightweight implementation of jQuery which provided access to the DOM on the server.
 
 ## Demo
 
@@ -26,20 +32,16 @@ This project was built with the following tools:
 
 I deployed the API using [Netlify](https://docs.netlify.com/functions/build-with-javascript/#format)'s serverless functions:
 
-- https://apotheken-notdienst-api.netlify.app/.netlify/functions/server
+* https://apotheken-notdienst-api.netlify.app/.netlify/functions/server
 
-- Query Parameters:
+* Query Parameters:
 
-  - `lat` | Latitude
-  - `lon` | Longitude
-  - `date` | UNIX Timestamp
+  + `lat` | Latitude
+  + `lon` | Longitude
+  + `date` | UNIX Timestamp
 
-- Example for Munich on 05.05.2020: https://apotheken-notdienst-api.netlify.app/.netlify/functions/server?lat=48.135124&lon=11.581981&date=1588702203721
+* Example for Munich on 05.05.2020: https://apotheken-notdienst-api.netlify.app/.netlify/functions/server?lat=48.135124&lon=11.581981&date=1588702203721
 
 ## License
 
 This project is licensed under the [MIT license](https://opensource.org/licenses/MIT).
-
-## Credits
-
-Thanks to [this](https://blog.logrocket.com/how-to-build-a-web-crawler-with-node/) tutorial by [Logrocket](https://logrocket.com).
